\name{funcClass.all}
\alias{funcClassPred}
\docType{data}
\title{
Classification methods used. Warning, in fscaret version 0.8.6.1 methods for variable importence ranking harvesting have not been implemented!
}
\description{
Vector of all classification methods used in solving problems by caret
}
\usage{data(funcClassPred)}
\format{
vector
}
\examples{

# Load library
library(fscaret)

# Load data set
data(funcClassPred)

# Print out object
funcClassPred


}
\keyword{datasets}
