\name{RMSE}
\alias{RMSE}
\title{
RMSE
}
\description{
Function calculates root mean squared error.
}
\usage{
RMSE(vect1, vect2, rows_no)
}
\arguments{
  \item{vect1}{
Numeric vector of predicted values
}
  \item{vect2}{
Numeric vector of observed values
}
  \item{rows_no}{
Number of observations
}
}
\author{
Aleksander Mendyk
}

\keyword{ math }
\keyword{ arith }
