\name{MSE}
\alias{MSE}
\title{
MSE
}
\description{
Function calculates mean squared error as predicted vs. observed
}
\usage{
MSE(vect1, vect2, rows_no)
}

\arguments{
  \item{vect1}{
Numeric vector of predicted values
}
  \item{vect2}{
Numeric vector of observed values
}
  \item{rows_no}{
Number of observations
}
}
\author{
Jakub Szlek and Aleksander Mendyk
}
\keyword{ math }
\keyword{ arith }
